addign read file to module api
