+++ Packs/Base/Scripts/DBotPreprocessTextData/DBotPreprocessTextData.py
@@ -25,34 +25,6 @@ HTML_PATTERNS = [
 html_parser = HTMLParser()
 
 
-def read_file(input_data, input_type):
-    data = []  # type: ignore
-    if not input_data:
-        return data
-    if input_type.endswith("string"):
-        if 'b64' in input_type:
-            input_data = base64.b64decode(input_data)
-            file_content = input_data.decode("utf-8")
-        else:
-            file_content = input_data
-    else:
-        res = demisto.getFilePath(input_data)
-        if not res:
-            return_error("Entry {} not found".format(input_data))
-        file_path = res['path']
-        if input_type.startswith('json'):
-            with open(file_path, 'r') as f:
-                file_content = f.read()
-    if input_type.startswith('csv'):
-        return pd.read_csv(file_path).fillna('').to_dict(orient='records')
-    elif input_type.startswith('json'):
-        return json.loads(file_content)
-    elif input_type.startswith('pickle'):
-        return pd.read_pickle(file_path, compression=None)
-    else:
-        return_error("Unsupported file type %s" % input_type)
-
-
 def concat_text_fields(data, target_field, text_fields):
     for d in data:
         text = ''
@@ -104,7 +76,7 @@ def pre_process(data, source_text_field, target_text_field, remove_html_tags, pr
 
 
 def pre_process_nlp(text_data):
-    tokenizer = Tokenizer(isValueJson=True, tokenizationMethod=demisto.args()['tokenizationMethod'],
+    tokenizer = Tokenizer(is_json=True, tokenization_method=demisto.args()['tokenizationMethod'],
                           language=demisto.args()['language'])
     processed_text_data = tokenizer.word_tokenize(json.dumps(text_data))
     if not isinstance(processed_text_data, list):
@@ -177,7 +149,7 @@ def main():
     output_original_text_fields = demisto.args().get('outputOriginalTextFields', 'false') == 'true'
     description = ""
     # read data
-    data = read_file(input, input_type)
+    data = FileReader.read_file(input, input_type)
 
     # concat text fields
     concat_text_fields(data, DBOT_TEXT_FIELD, text_fields)
@@ -231,10 +203,7 @@ def main():
     }
     return entry
 
-try:
-    from MLApiModule import *  # noqa: E402
-except Exception as e:
-    demisto.log('could not import MLApiModule')
+from MLApiModule import *  # noqa: E402
 
 if __name__ in ['builtins', '__main__']:
     entry = main()
